# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Federation API
  version: 0.1.0
  description: |
    Backend REST, ActivityPub, C2S, S2S inbox, WebFinger, and delivery administration contract.

    Bounded REST list operations return JSON arrays for backward compatibility and expose a uniform
    offset-pagination contract through `Link`, `X-Pagination-Limit`, `X-Pagination-Offset`,
    `X-Pagination-Count`, and `X-Pagination-Has-More` response headers. The `Link` header preserves
    active filters and supplies RFC 8288 `prev` and `next` relations when applicable.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/projects/archived:
    get:
      tags: [Projects]
      operationId: listArchivedProjects
      security: [{cookieAuth: []}]
      responses:
        "200":
          description: Restorable projects for the current user.
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/ArchivedProject"}}
  /api/v1/projects/{projectID}/activity:
    get:
      tags: [Projects]
      operationId: listProjectActivity
      security: [{cookieAuth: []}]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Immutable project and ticket change history.
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/ProjectActivityEvent"}}
  /api/v1/projects/{projectID}/tickets/archived:
    get:
      tags: [Tickets]
      operationId: listArchivedTickets
      security: [{cookieAuth: []}]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Restorable project tickets.
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/ArchivedTicket"}}
  /api/v1/projects/{projectID}/archive:
    post:
      tags: [Projects]
      operationId: archiveProject
      security: [{cookieAuth: []}]
      parameters: [{$ref: "#/components/parameters/ProjectID"}, {$ref: "#/components/parameters/IfMatch"}]
      responses: {"204": {description: Project archived.}, "412": {$ref: "#/components/responses/Error"}, "428": {$ref: "#/components/responses/Error"}}
  /api/v1/projects/{projectID}/restore:
    post:
      tags: [Projects]
      operationId: restoreProject
      security: [{cookieAuth: []}]
      parameters: [{$ref: "#/components/parameters/ProjectID"}, {$ref: "#/components/parameters/IfMatch"}]
      responses: {"204": {description: Project restored.}, "412": {$ref: "#/components/responses/Error"}, "428": {$ref: "#/components/responses/Error"}}
  /api/v1/tickets/{ticketID}/archive:
    post:
      tags: [Tickets]
      operationId: archiveTicket
      security: [{cookieAuth: []}]
      parameters: [{$ref: "#/components/parameters/TicketID"}, {$ref: "#/components/parameters/IfMatch"}]
      responses: {"204": {description: Ticket archived.}, "412": {$ref: "#/components/responses/Error"}, "428": {$ref: "#/components/responses/Error"}}
  /api/v1/tickets/{ticketID}/restore:
    post:
      tags: [Tickets]
      operationId: restoreTicket
      security: [{cookieAuth: []}]
      parameters: [{$ref: "#/components/parameters/TicketID"}, {$ref: "#/components/parameters/IfMatch"}]
      responses: {"204": {description: Ticket restored.}, "412": {$ref: "#/components/responses/Error"}, "428": {$ref: "#/components/responses/Error"}}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: progo.session
      description: HttpOnly SameSite=Strict browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsBearerAuth:
      type: http
      scheme: bearer
      description: Optional in development; required when METRICS_TOKEN is configured.
...
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong numeric entity version returned as ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TicketID:
      name: ticketID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    ArchivedProject:
      type: object
      required: [id, name, description, version, archived_at]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        description: {type: string}
        version: {type: integer, format: int64, minimum: 1}
        archived_at: {type: string, format: date-time}
...
    ArchivedTicket:
      type: object
      required: [id, project_id, title, version, archived_at]
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        title: {type: string}
        version: {type: integer, format: int64, minimum: 1}
        archived_at: {type: string, format: date-time}
    ProjectActivityEvent:
      type: object
      required: [id, project_id, entity_type, entity_id, action, created_at]
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        actor_id: {type: [string, "null"], format: uuid}
        actor_handle: {type: [string, "null"]}
        entity_type: {type: string, enum: [project, ticket]}
        entity_id: {type: string, format: uuid}
        action: {type: string, enum: [created, updated, archived, restored]}
        before_state: {type: [object, "null"]}
        after_state: {type: [object, "null"]}
        created_at: {type: string, format: date-time}
...